requirements.txt
tqdm
matplotlib
seaborn
scikit-learn
numpy
ipykernel
torch
torchvision
streamlit
Pillow